// apps/web/src/app/features/inventory/modules/budget-categories/types/budget-categories.types.ts
export interface BudgetCategorie {
  id: number;
  category_code: string;
  category_name: string;
  accounting_code?: string | null;
  description?: string | null;
  is_active?: boolean | null;
  created_at?: string | null;
  updated_at?: string | null;
}